backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-multipart==0.0.12
docker>=7.1.0
pynvml==11.5.0
psutil==6.0.0
pydantic==2.9.0
huggingface-hub==0.24.0
pyyaml==6.0.1
websockets>=12.0
passlib[bcrypt]==1.7.4
bcrypt<5.0.0
python-jose[cryptography]==3.3.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
pytest==8.3.0
httpx>=0.27.0
slowapi==0.1.9
redis==5.2.1
